feat: modify settings page, add account details and preference components
